Added explicit explicit type in Berkeley DB flag definitions * Unless specific flag types are added, building with Berkeley DB   v18.1.25 produces compiler errors reporting ambiguous calls to   relevant methods.

+++ src/berkeleydb_tmpl.cpp
@@ -30,8 +30,8 @@
 //
 //
 //
-#define DBFLAGS               0
-#define DBENVFLAGS            0
+#define DBFLAGS               ((u_int32_t) 0)
+#define DBENVFLAGS            ((u_int32_t) 0)
 
 #define FILEMASK              0664              // database file access mask (rw-rw-r--)
